Compile generated erlang during dictionary file compilation
Some dictionaries depend on other dictionaries being processed. In order for the dependency to resolve, the generated erl file needs to be compiled and the diameter compiler needs to be pointed to compiled file. This change adds that compilation. In addition, it includes passing the dia_opts down to the dictionary processor. It works more like the original diameter compiler in rebar 2.