fix compiler invocation
The first argument to rebar_base_compiler:run needs to be the rebar
config.

We used to get away with passing something else in, but recent change
in rebars core changed that. We now have to pass a valid config.

Use a wrapper fun to pass the arguments we need into compile_dia. And
since we already get the config, remove the rebar state and use the
config directly.